친구 기능 예외 처리 (#57)
* 친구 기능 예외 처리

* 친구 기능 예외 처리

+++ src/main/java/com/beyond/sportsmatch/domain/friend/model/service/FriendRequestServiceImpl.java
@@ -1,6 +1,8 @@
 package com.beyond.sportsmatch.domain.friend.model.service;
 
 
+import com.beyond.sportsmatch.common.exception.SportsMatchException;
+import com.beyond.sportsmatch.common.exception.message.ExceptionMessage;
 import com.beyond.sportsmatch.domain.friend.model.dto.FriendResponseDto;
 import com.beyond.sportsmatch.domain.friend.model.entity.Friend;
 import com.beyond.sportsmatch.domain.friend.model.entity.FriendRequest;
@@ -35,11 +37,22 @@ public class FriendRequestServiceImpl implements FriendRequestService {
 
     @Override
     public void deleteSentFriendRequest(int senderUserId, int receiverUserId) {
+        boolean exists = friendRequestRepository.existsBySenderUserIdAndReceiverUserId(senderUserId, receiverUserId);
+
+        if (!exists) {
+            throw new SportsMatchException(ExceptionMessage.REQUEST_NOT_FOUND);
+        }
+
         friendRequestRepository.deleteBySenderUserIdAndReceiverUserId(senderUserId, receiverUserId);
     }
 
     @Override
     public void deleteReceivedFriendRequest(int receiverUserId, int senderUserId) {
+        boolean exists = friendRequestRepository.existsBySenderUserIdAndReceiverUserId(receiverUserId, senderUserId);
+
+        if (!exists) {
+            throw new SportsMatchException(ExceptionMessage.REQUEST_NOT_FOUND);
+        }
         friendRequestRepository.deleteByReceiverUserIdAndSenderUserId(receiverUserId, senderUserId);
     }
 
@@ -47,13 +60,13 @@ public class FriendRequestServiceImpl implements FriendRequestService {
     @Transactional
     public void sendFriendRequest(int senderUserId, String receiverUserNickname) {
         User sender = userRepository.findById(senderUserId)
-                .orElseThrow(() -> new RuntimeException("보낸 사람 없음"));
+                .orElseThrow(() -> new SportsMatchException(ExceptionMessage.USER_NOT_FOUND));
         User receiver = userRepository.findByNickname(receiverUserNickname)
-                .orElseThrow(() -> new RuntimeException("받는 사람 없음"));
+                .orElseThrow(() -> new SportsMatchException(ExceptionMessage.USER_NOT_FOUND));
 
         boolean exists = friendRequestRepository.existsBySenderUserIdAndReceiverUserIdAndStatus(sender, receiver, "Pending");
         if (exists) {
-            throw new RuntimeException("이미 요청한 친구입니다.");
+            throw new SportsMatchException(ExceptionMessage.ALREADY_SENT_REQUEST);
         }
 
 
@@ -66,13 +79,22 @@ public class FriendRequestServiceImpl implements FriendRequestService {
 
     @Override
     public void acceptFriendRequest(int senderUserId, int receiverUserId) {
+        // 친구인지 먼저 확인
+        boolean alreadyFriend = friendRepository.existsByLoginUserIdUserIdAndFriendUserIdUserId(senderUserId, receiverUserId);
+        if (alreadyFriend) {
+            throw new SportsMatchException(ExceptionMessage.ALREADY_FRIEND);
+        }
+
+
         friendRequestRepository.deleteByReceiverUserIdAndSenderUserId(receiverUserId, senderUserId);
+
+
 
         // 2. Friend 엔티티에 양방향 저장
         User sender = userRepository.findById(senderUserId)
-                .orElseThrow(() -> new IllegalArgumentException("보낸 유저를 찾을 수 없습니다."));
+                .orElseThrow(() -> new SportsMatchException(ExceptionMessage.USER_NOT_FOUND));
         User receiver = userRepository.findById(receiverUserId)
-                .orElseThrow(() -> new IllegalArgumentException("받는 유저를 찾을 수 없습니다."));
+                .orElseThrow(() -> new SportsMatchException(ExceptionMessage.USER_NOT_FOUND));
 
         Friend friend1 = new Friend();
         friend1.setLoginUserId(sender);
